Resolved conflicts in batchgen/sequence.py:
- __slots__: combined pool fields (batch_id, pool_slot_index, priority)
  with main's additions (lifespan monitoring, repetition detection)
- __init__: combined both sets of field initializations + log_event()

New from main:
- FP32 logits, repetition detection, host KV page fix (PR #122)
- N-gram repetition detection improvements (PR #123)
- NCCL tree, batch-invariant gate/lm_head, MoE padding mask (PR #124/#125)
- MiniMax-M25 FP8 blockwise MoE + packed QKV (PR #115, #119, #121)
- Dead code removal + test consolidation (PR #114)

+++ batchgen/models/openai/gpt_oss_120b/model.py
@@ -1742,7 +1742,7 @@ class GptOss(nn.Module):
         )
 
         hidden_states = outputs[0]
-        logits = self.lm_head(hidden_states)
+        logits = self.lm_head(hidden_states).float()
 
         # Debug logging for logits analysis
         if os.environ.get("BATCHGEN_DEBUG_LOGITS", "0") == "1":
